create utility function for creating new schedule docs
