+++ src/hooks/sections/useCreateSection.ts
@@ -2,7 +2,7 @@ import { db } from "@/config/firebase";
 import { createSectionDoc } from "@/data/firestore/sections";
 import { setSectionScheduleDoc } from "@/data/firestore/sectionSchedules";
 import { SectionScheduleDoc } from "@/data/firestore/types/documents";
-import { getBlockIdFromNum } from "@/types/scheduling/schedulingUtils";
+import { DEFAULT_NUMBER_BLOCKS, getBlockIdFromNum, getEmptySectionScheduleDoc } from "@/types/scheduling/schedulingUtils";
 import { SectionType } from "@/types/sessions/sessionTypes";
 import { useMutation } from "@tanstack/react-query";
 import { runTransaction, Timestamp, Transaction } from "firebase/firestore";
@@ -16,8 +16,6 @@ interface CreateSectionRequest {
   type: SectionType;
 }
 
-const DEFAULT_NUMBER_BLOCKS = 5;
-
 async function createSection(req: CreateSectionRequest) {
   const { sessionId, ...rest } = req;
   if (rest.type === "COMMON") {
@@ -30,17 +28,7 @@ async function createSection(req: CreateSectionRequest) {
     return;
   }
 
-  const sectionScheduleDoc: SectionScheduleDoc = {
-    type: rest.type,
-    blocks: {},
-    alternatePeriodsOff: {}
-  };
-  for (let i = 0; i < DEFAULT_NUMBER_BLOCKS; i++) {
-    sectionScheduleDoc.blocks[getBlockIdFromNum(i)] = {
-      activities: [],
-      periodsOff: []
-    }
-  }
+  const sectionScheduleDoc = getEmptySectionScheduleDoc(rest.type, DEFAULT_NUMBER_BLOCKS);
   await runTransaction(db, async (transaction: Transaction) => {
     const sectionId = await createSectionDoc(sessionId, {
       name: rest.name,
